CombustiblePotato/requirements.txt
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.1.8
coverage==7.7.0
cryptography==44.0.2
flexcache==0.3
flexparser==0.4
iniconfig==2.0.0
mypy-extensions==1.0.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
phoenix6==25.1.0
photonlibpy==2025.1.1
Pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pyfrc==2025.1.0
PyNaCl==1.5.0
pynetconsole==2.0.4
pyntcore==2025.3.1.1
pytest==8.3.5
pytest-reraise==2.1.2
robotpy==2025.3.1.1
robotpy-apriltag==2025.3.1.1
robotpy-cli==2024.0.0
robotpy-commands-v2==2025.1.1
robotpy-cscore==2025.3.1.1
robotpy-hal==2025.3.1.1
robotpy-halsim-gui==2025.3.1.1
robotpy-installer==2025.1.0
robotpy-navx==2025.0.1
robotpy-rev==2025.0.2
robotpy-wpilib-utilities==2025.0.0
robotpy-wpimath==2025.3.1.1
robotpy-wpinet==2025.3.1.1
robotpy-wpiutil==2025.3.1.1
setuptools==76.0.0
sleipnirgroup-choreolib==2025.0.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
wpilib==2025.3.1.1
robotpy-pathplannerlib==2025.2.6
